added CRUD controller implementations and update routing

+++ zadanie2/play-crud/app/models/Cart.scala
@@ -2,7 +2,13 @@ package models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Cart(id: Long, customer: Long, productsQuantity: Int, totalProductsPrice: BigDecimal, coupon: Long, createdAt: String)
+case class Cart(id: Int,
+                customer: Int,
+                productsQuantity: Int,
+                totalProductsPrice: BigDecimal,
+                coupon: Int,
+                createdAt: String
+               )
 
 object Cart {
     implicit val cartFormat: OFormat[Cart] = Json.format[Cart]
