+++ zadanie2/play-crud/app/models/CartItem.scala
@@ -2,7 +2,11 @@ package models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class CartItem(id: Long, cart: Long, product: Long, productQuantity: Int)
+case class CartItem(id: Int,
+                    cart: Int,
+                    product: Int,
+                    productQuantity: Int
+                   )
 
 object CartItem {
     implicit val cartItemFormat: OFormat[CartItem] = Json.format[CartItem]
