+++ zadanie2/play-crud/app/models/Coupon.scala
@@ -2,7 +2,14 @@ package models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Coupon(id: Long, code: Long, couponType: String, discount: BigDecimal, isActive: Boolean, createdAt: String, usedAt: String)
+case class Coupon(id: Int,
+                  code: Long,
+                  couponType: String,
+                  discount: BigDecimal,
+                  isActive: Boolean,
+                  createdAt: String,
+                  usedAt: String
+                 )
 
 object Coupon {
     implicit val couponFormat: OFormat[Coupon] = Json.format[Coupon]
