+++ zadanie2/play-crud/app/models/OrderItem.scala
@@ -2,7 +2,10 @@ package models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class OrderItem(id: Long, order: Long, product: Long)
+case class OrderItem(id: Int,
+                     order: Int,
+                     product: Int
+                    )
 
 object OrderItem {
     implicit val orderItemFormat: OFormat[OrderItem] = Json.format[OrderItem]
