+++ zadanie2/play-crud/app/models/Product.scala
@@ -2,7 +2,12 @@ package models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Product(id: Long, name: String, description: String, price: Int, category: Int)
+case class Product(id: Int,
+                   name: String,
+                   description: String,
+                   price: Float,
+                   category: Int
+                  )
 
 object Product {
     implicit val productFormat: OFormat[Product] = Json.format[Product]
