+++ zadanie2/play-crud/app/models/Review.scala
@@ -2,7 +2,12 @@ package models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Review(id: Int, product: Long, customer: Long, content: String, score: Float)
+case class Review(id: Int,
+                  product: Int,
+                  customer: Int,
+                  content: String,
+                  score: Float
+                 )
 
 object Review {
     implicit val reviewFormat: OFormat[Review] = Json.format[Review]
