+++ zadanie2/play-crud/app/models/repository/OrderRepository.scala
@@ -17,14 +17,14 @@ class OrderRepository @Inject()(dbConfigProvider: DatabaseConfigProvider,
     import profile.api._
 
     class OrderTable(tag: Tag) extends Table[Order](tag, "order") {
-        def id = column[Long]("id", O.PrimaryKey, O.AutoInc)
+        def id = column[Int]("id", O.PrimaryKey, O.AutoInc)
         def createdAt = column[String]("createdAt")
-        def customer = column[Long]("customer")
+        def customer = column[Int]("customer")
         def customer_fk = foreignKey("customer_fk", customer, customerVal)(_.id)
         def isPaid = column[Boolean]("isPaid")
         def paidAt = column[String]("paidAt")
         def totalOrderValue = column[Int]("totalOrderValue")
-        def coupon = column[Long]("coupon")
+        def coupon = column[Int]("coupon")
         def coupon_fk = foreignKey("coupon_fk", coupon, couponVal)(_.id)
 
         def * = (id, createdAt, customer, isPaid, paidAt, totalOrderValue, coupon) <> ((Order.apply _).tupled, Order.unapply)
@@ -37,7 +37,7 @@ class OrderRepository @Inject()(dbConfigProvider: DatabaseConfigProvider,
     val customerVal = TableQuery[CustomerTable]
     val couponVal = TableQuery[CouponTable]
 
-    def create(createdAt: String, customer: Long, isPaid: Boolean, paidAt: String, totalOrderValue: Int, coupon: Long): Future[Order] = db.run {
+    def create(createdAt: String, customer: Int, isPaid: Boolean, paidAt: String, totalOrderValue: Int, coupon: Int): Future[Order] = db.run {
         (order.map(o => (o.createdAt, o.customer, o.isPaid, o.paidAt, o.totalOrderValue, o.coupon))
             returning order.map(_.id)
             into {case((createdAt, customer, isPaid, paidAt, totalOrderValue, coupon), id) => Order(id, createdAt, customer, isPaid, paidAt, totalOrderValue, coupon)}
@@ -48,11 +48,11 @@ class OrderRepository @Inject()(dbConfigProvider: DatabaseConfigProvider,
         order.result
     }
 
-    def getById(id: Long): Future[Order] = db.run {
+    def getById(id: Int): Future[Order] = db.run {
         order.filter(_.id === id).result.head
     }
 
-    def getByCustomer(customer_id: Long): Future[Seq[Order]] = db.run {
+    def getByCustomer(customer_id: Int): Future[Seq[Order]] = db.run {
         order.filter(_.customer === customer_id).result
     }
